pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visionplant"
version = "1.0.0"
description = "API profesional de reconocimiento de plantas con IA avanzada"
authors = [
    {name = "VisionPlant Team", email = "team@visionplant.dev"}
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "torch>=2.1.1",
    "torchvision>=0.16.1",
    "numpy>=1.24.3",
    "Pillow>=10.1.0",
    "pydantic>=2.5.0",
]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".venv"]
